+++ e2e-tests/development-runtime/scripts/cms-webhook.mjs
@@ -1,0 +1,11 @@
+import fetch from "node-fetch"
+
+fetch(`http://localhost:8000/__refresh/${process.argv[2]}`, {
+  method: `POST`,
+  headers: {
+    "Content-Type": `application/json`,
+  },
+  body: JSON.stringify({
+    "fake-data-update": true,
+  }),
+})