+++ e2e-tests/production-runtime/cypress/support/e2e.ts
@@ -82,7 +82,6 @@ Cypress.Commands.add(
       win.___navigate(pathname)
     })
 
-    // @ts-expect-error - gatsby-cypress doesn't have types
     return cy.waitForAPI(`onRouteUpdate`).then(() => subject)
   }
 )