+++ examples/using-cypress/cypress.config.js
@@ -1,8 +1,0 @@
-const { defineConfig } = require('cypress')
-
-module.exports = defineConfig({
-  e2e: {
-    baseUrl: 'http://localhost:8000',
-    specPattern: "cypress/e2e"
-  }
-})