+++ examples/using-cypress/cypress/support/e2e.js
@@ -1,3 +1,0 @@
-import "./commands"
-import "cypress-axe"
-import "@testing-library/cypress/add-commands"