+++ examples/using-cypress/cypress/support/e2e.ts
@@ -1,0 +1,3 @@
+import "gatsby-cypress/commands"
+import "cypress-axe"
+import "@testing-library/cypress/add-commands"