+++ examples/using-drupal/gatsby-config.js
@@ -8,6 +8,7 @@ module.exports = {
       options: {
         baseUrl: `https://live-contentacms.pantheonsite.io/`,
         apiBase: `api`,
+        typePrefix: "Drupal",
       },
     },
     {