+++ integration-tests/functions/src/pages/index.js
@@ -82,7 +82,7 @@ const badgeStyle = {
 const links = [
   {
     text: "Tutorial",
-    url: "https://www.gatsbyjs.com/docs/tutorial/",
+    url: "https://www.gatsbyjs.com/docs/tutorial/getting-started/",
     description:
       "A great place to get started if you're new to web development. Designed to guide you through setting up your first Gatsby site.",
     color: "#E95800",