+++ packages/gatsby-source-contentful/src/source-nodes.js
@@ -1,7 +1,12 @@
 // @ts-check
-import { hasFeature } from "gatsby-plugin-utils/has-feature"
 import isOnline from "is-online"
 import _ from "lodash"
+import {
+  addNodeToExistingNodesCache,
+  getExistingCachedNodes,
+  removeNodeFromExistingNodesCache,
+} from "./backreferences"
+import { untilNextEventLoopTick } from "./utils"
 
 import { downloadContentfulAssets } from "./download-contentful-assets"
 import { fetchContent } from "./fetch"
@@ -12,10 +17,10 @@ import {
   createAssetNodes,
   createNodesForContentType,
   makeId,
-  makeTypeName,
 } from "./normalize"
 import { createPluginConfig } from "./plugin-options"
 import { CODES } from "./report"
+import { hasFeature } from "gatsby-plugin-utils/has-feature"
 
 const conflictFieldPrefix = `contentful`
 
@@ -41,13 +46,10 @@ const CONTENT_DIGEST_COUNTER_SEPARATOR = `_COUNT_`
  * possible for each localized node i.e. get the localized field if it exists
  * or the fallback field or the default field.
  */
-
-let isFirstSourceNodesCallOfCurrentNodeProcess = true
 export async function sourceNodes(
   {
     actions,
     getNode,
-    getNodes,
     createNodeId,
     store,
     cache,
@@ -57,61 +59,45 @@ export async function sourceNodes(
   },
   pluginOptions
 ) {
-  const hasStatefulSourceNodes = hasFeature(`stateful-source-nodes`)
-  const needToTouchNodes = !hasStatefulSourceNodes
-
   const {
-    createNode,
+    createNode: originalCreateNode,
     touchNode,
-    deleteNode,
+    deleteNode: originalDeleteNode,
     unstable_createNodeManifest,
     enableStatefulSourceNodes,
   } = actions
 
-  const online = await isOnline()
-
-  if (hasStatefulSourceNodes) {
+  if (hasFeature(`stateful-source-nodes`)) {
     enableStatefulSourceNodes()
-  }
-  // Gatsby only checks if a node has been touched on the first sourcing.
-  // As iterating and touching nodes can grow quite expensive on larger sites with
-  // 1000s of nodes, we'll skip doing this on subsequent sources.
-  else if (isFirstSourceNodesCallOfCurrentNodeProcess && needToTouchNodes) {
-    getNodes().forEach(node => {
-      if (node.internal.owner !== `gatsby-source-contentful`) {
-        return
-      }
-      touchNode(node)
-      if (node?.fields?.localFile) {
-        // Prevent GraphQL type inference from crashing on this property
-        touchNode(getNode(node.fields.localFile))
-      }
-    })
-  }
-
-  isFirstSourceNodesCallOfCurrentNodeProcess = false
-
-  if (
-    !online &&
-    process.env.GATSBY_CONTENTFUL_OFFLINE === `true` &&
-    process.env.NODE_ENV !== `production`
-  ) {
-    return
   }
 
   const pluginConfig = createPluginConfig(pluginOptions)
-  const sourceId = `${pluginConfig.get(`spaceId`)}-${pluginConfig.get(
-    `environment`
-  )}`
 
-  const fetchActivity = reporter.activityTimer(`Contentful: Fetch data`, {
-    parentSpan,
-  })
+  // wrap createNode so we can cache them in memory for faster lookups when finding backreferences
+  const createNode = node => {
+    addNodeToExistingNodesCache(node)
+
+    return originalCreateNode(node)
+  }
+
+  const deleteNode = node => {
+    removeNodeFromExistingNodesCache(node)
+
+    return originalDeleteNode(node)
+  }
+
+  // Array of all existing Contentful nodes
+  const { existingNodes, memoryNodeCountsBySysType } =
+    await getExistingCachedNodes({
+      actions,
+      getNode,
+      pluginConfig,
+    })
 
   // If the user knows they are offline, serve them cached result
   // For prod builds though always fail if we can't get the latest data
   if (
-    !online &&
+    !(await isOnline()) &&
     process.env.GATSBY_CONTENTFUL_OFFLINE === `true` &&
     process.env.NODE_ENV !== `production`
   ) {
@@ -121,12 +107,19 @@ export async function sourceNodes(
     )
 
     return
-  }
-  if (process.env.GATSBY_CONTENTFUL_OFFLINE) {
+  } else if (process.env.GATSBY_CONTENTFUL_OFFLINE) {
     reporter.info(
       `Note: \`GATSBY_CONTENTFUL_OFFLINE\` was set but it either was not \`true\`, we _are_ online, or we are in production mode, so the flag is ignored.`
     )
   }
+
+  const sourceId = `${pluginConfig.get(`spaceId`)}-${pluginConfig.get(
+    `environment`
+  )}`
+
+  const fetchActivity = reporter.activityTimer(`Contentful: Fetch data`, {
+    parentSpan,
+  })
 
   fetchActivity.start()
 
@@ -148,13 +141,14 @@ export async function sourceNodes(
     store.getState().status.plugins?.[`gatsby-source-contentful`]?.[
       CACHE_SYNC_TOKEN
     ]
+  const isCachedBuild = !!syncToken
 
   // Actual fetch of data from Contentful
   const {
     currentSyncData,
     tagItems,
     defaultLocale,
-    locales: allLocales,
+    locales: allLocales = [],
     space,
   } = await fetchContent({ syncToken, pluginConfig, reporter })
 
@@ -185,7 +179,7 @@ export async function sourceNodes(
   }
 
   // Update syncToken
-  const nextSyncToken = currentSyncData.nextSyncToken
+  const nextSyncToken = currentSyncData?.nextSyncToken
 
   actions.setPluginStatus({
     [CACHE_SYNC_TOKEN]: nextSyncToken,
@@ -202,31 +196,20 @@ export async function sourceNodes(
   )
   processingActivity.start()
 
-  // Array of all existing Contentful nodes
-  const existingNodes = getNodes().filter(
-    n =>
-      n.internal.owner === `gatsby-source-contentful` &&
-      (pluginConfig.get(`enableTags`)
-        ? n.internal.type !== `ContentfulTag`
-        : true)
-  )
-
   // Report existing, new and updated nodes
   const nodeCounts = {
     newEntry: 0,
     newAsset: 0,
     updatedEntry: 0,
     updatedAsset: 0,
-    existingEntry: 0,
-    existingAsset: 0,
-    deletedEntry: currentSyncData.deletedEntries.length,
-    deletedAsset: currentSyncData.deletedAssets.length,
+    deletedEntry: currentSyncData?.deletedEntries?.length || 0,
+    deletedAsset: currentSyncData?.deletedAssets?.length || 0,
   }
-  existingNodes.forEach(node => nodeCounts[`existing${node.sys.type}`]++)
-  currentSyncData.entries.forEach(entry =>
+
+  currentSyncData?.entries?.forEach(entry =>
     entry.sys.revision === 1 ? nodeCounts.newEntry++ : nodeCounts.updatedEntry++
   )
-  currentSyncData.assets.forEach(asset =>
+  currentSyncData?.assets?.forEach(asset =>
     asset.sys.revision === 1 ? nodeCounts.newAsset++ : nodeCounts.updatedAsset++
   )
 
@@ -234,12 +217,16 @@ export async function sourceNodes(
   reporter.info(`Contentful: ${nodeCounts.updatedEntry} updated entries`)
   reporter.info(`Contentful: ${nodeCounts.deletedEntry} deleted entries`)
   reporter.info(
-    `Contentful: ${nodeCounts.existingEntry / locales.length} cached entries`
+    `Contentful: ${
+      memoryNodeCountsBySysType.Entry / locales.length
+    } cached entries`
   )
   reporter.info(`Contentful: ${nodeCounts.newAsset} new assets`)
   reporter.info(`Contentful: ${nodeCounts.updatedAsset} updated assets`)
   reporter.info(
-    `Contentful: ${nodeCounts.existingAsset / locales.length} cached assets`
+    `Contentful: ${
+      memoryNodeCountsBySysType.Asset / locales.length
+    } cached assets`
   )
   reporter.info(`Contentful: ${nodeCounts.deletedAsset} deleted assets`)
 
@@ -276,7 +263,7 @@ export async function sourceNodes(
 
   reporter.verbose(`Resolving Contentful references`)
 
-  const newOrUpdatedEntries = new Set()
+  let newOrUpdatedEntries = new Set()
   entryList.forEach(entries => {
     entries.forEach(entry => {
       newOrUpdatedEntries.add(`${entry.sys.id}___${entry.sys.type}`)
@@ -329,15 +316,20 @@ export async function sourceNodes(
   }
 
   // Update existing entry nodes that weren't updated but that need reverse links added or removed.
-  const existingNodesThatNeedReverseLinksUpdateInDatastore = new Set()
-  existingNodes
-    .filter(
-      n =>
-        n.sys.type === `Entry` &&
-        !newOrUpdatedEntries.has(`${n.id}___${n.sys.type}`) &&
-        !deletedEntryGatsbyReferenceIds.has(n.id)
-    )
-    .forEach(n => {
+  let existingNodesThatNeedReverseLinksUpdateInDatastore = new Set()
+
+  if (isCachedBuild) {
+    existingNodes.forEach(n => {
+      if (
+        !(
+          n.sys.type === `Entry` &&
+          !newOrUpdatedEntries.has(`${n.id}___${n.sys.type}`) &&
+          !deletedEntryGatsbyReferenceIds.has(n.id)
+        )
+      ) {
+        return
+      }
+
       if (
         n.contentful_id &&
         foreignReferenceMap[`${n.contentful_id}___${n.sys.type}`]
@@ -398,10 +390,17 @@ export async function sourceNodes(
         })
       }
     })
+  }
+
+  // allow node to gc if it needs to
+  // @ts-ignore
+  newOrUpdatedEntries = undefined
+  await untilNextEventLoopTick()
 
   // We need to call `createNode` on nodes we modified reverse links on,
   // otherwise changes won't actually persist
   if (existingNodesThatNeedReverseLinksUpdateInDatastore.size) {
+    let existingNodesLoopCount = 0
     for (const node of existingNodesThatNeedReverseLinksUpdateInDatastore) {
       function addChildrenToList(node, nodeList = [node]) {
         for (const childNodeId of node?.children ?? []) {
@@ -422,20 +421,10 @@ export async function sourceNodes(
         // We should not mutate original node as Gatsby will still
         // compare against what's in in-memory weak cache, so we
         // clone original node to ensure reference identity is not possible
-        const nodeToUpdate = _.cloneDeep(nodeToUpdateOriginal)
-        // We need to remove properties from existing fields
-        // that are reserved and managed by Gatsby (`.internal.owner`, `.fields`).
-        // Gatsby automatically will set `.owner` it back
-        delete nodeToUpdate.internal.owner
-        // `.fields` need to be created with `createNodeField` action, we can't just re-add them.
-        // Other plugins (or site itself) will have opportunity to re-generate them in `onCreateNode` lifecycle.
-        // Contentful content nodes are not using `createNodeField` so it's safe to delete them.
-        // (Asset nodes DO use `createNodeField` for `localFile` and if we were updating those, then
-        // we would also need to restore that field ourselves after re-creating a node)
-        delete nodeToUpdate.fields // plugin adds node field on asset nodes which don't have reverse links
+        const nodeToUpdate = nodeToUpdateOriginal.__memcache
+          ? getNode(nodeToUpdateOriginal.id)
+          : nodeToUpdateOriginal
 
-        // We add or modify counter postfix to contentDigest
-        // to make sure Gatsby treat this as data update
         let counter
         const [initialContentDigest, counterStr] =
           nodeToUpdate.internal.contentDigest.split(
@@ -452,11 +441,52 @@ export async function sourceNodes(
           counter++
         }
 
-        nodeToUpdate.internal.contentDigest = `${initialContentDigest}${CONTENT_DIGEST_COUNTER_SEPARATOR}${counter}`
-        createNode(nodeToUpdate)
+        const newNode = {
+          ...nodeToUpdate,
+          internal: {
+            ...nodeToUpdate.internal,
+            // We need to remove properties from existing fields
+            // that are reserved and managed by Gatsby (`.internal.owner`, `.fields`).
+            // Gatsby automatically will set `.owner` it back
+            owner: undefined,
+            // We add or modify counter postfix to contentDigest
+            // to make sure Gatsby treat this as data update
+            contentDigest: `${initialContentDigest}${CONTENT_DIGEST_COUNTER_SEPARATOR}${counter}`,
+          },
+          // `.fields` need to be created with `createNodeField` action, we can't just re-add them.
+          // Other plugins (or site itself) will have opportunity to re-generate them in `onCreateNode` lifecycle.
+          // Contentful content nodes are not using `createNodeField` so it's safe to delete them.
+          // (Asset nodes DO use `createNodeField` for `localFile` and if we were updating those, then
+          // we would also need to restore that field ourselves after re-creating a node)
+          fields: undefined, // plugin adds node field on asset nodes which don't have reverse links
+        }
+
+        // memory cached nodes are mutated during back reference checks
+        // so we need to carry over the changes to the updated node
+        if (node.__memcache) {
+          for (const key of Object.keys(node)) {
+            if (!key.endsWith(`___NODE`)) {
+              continue
+            }
+
+            newNode[key] = node[key]
+          }
+        }
+
+        createNode(newNode)
+
+        if (existingNodesLoopCount++ % 2000 === 0) {
+          // dont block the event loop
+          await untilNextEventLoopTick()
+        }
       }
     }
   }
+
+  // allow node to gc if it needs to
+  // @ts-ignore
+  existingNodesThatNeedReverseLinksUpdateInDatastore = undefined
+  await untilNextEventLoopTick()
 
   const creationActivity = reporter.activityTimer(`Contentful: Create nodes`, {
     parentSpan,
@@ -478,26 +508,28 @@ export async function sourceNodes(
 
     // A contentType can hold lots of entries which create nodes
     // We wait until all nodes are created and processed until we handle the next one
-    // TODO add batching in gatsby-core
-    await Promise.all(
-      createNodesForContentType({
-        contentTypeItem,
-        restrictedNodeFields,
-        conflictFieldPrefix,
-        entries: entryList[i],
-        createNode,
-        createNodeId,
-        getNode,
-        resolvable,
-        foreignReferenceMap,
-        defaultLocale,
-        locales,
-        space,
-        useNameForId: pluginConfig.get(`useNameForId`),
-        pluginConfig,
-        unstable_createNodeManifest,
-      })
-    )
+    await createNodesForContentType({
+      contentTypeItem,
+      restrictedNodeFields,
+      conflictFieldPrefix,
+      entries: entryList[i],
+      createNode,
+      createNodeId,
+      getNode,
+      resolvable,
+      foreignReferenceMap,
+      defaultLocale,
+      locales,
+      space,
+      useNameForId: pluginConfig.get(`useNameForId`),
+      pluginConfig,
+      unstable_createNodeManifest,
+    })
+
+    // allow node to garbage collect these items if it needs to
+    contentTypeItems[i] = undefined
+    entryList[i] = undefined
+    await untilNextEventLoopTick()
   }
 
   if (assets.length) {
@@ -508,19 +540,24 @@ export async function sourceNodes(
   for (let i = 0; i < assets.length; i++) {
     // We wait for each asset to be process until handling the next one.
     assetNodes.push(
-      ...(await Promise.all(
-        createAssetNodes({
-          assetItem: assets[i],
-          createNode,
-          createNodeId,
-          defaultLocale,
-          locales,
-          space,
-          pluginConfig,
-        })
-      ))
+      ...(await createAssetNodes({
+        assetItem: assets[i],
+        createNode,
+        createNodeId,
+        defaultLocale,
+        locales,
+        space,
+        pluginConfig,
+      }))
     )
+
+    assets[i] = undefined
+    if (i % 1000 === 0) {
+      await untilNextEventLoopTick()
+    }
   }
+
+  await untilNextEventLoopTick()
 
   // Create tags entities
   if (tagItems.length) {