+++ packages/gatsby-source-drupal/src/gatsby-node.ts
@@ -21,17 +21,17 @@ import {
   downloadFile,
   isFileNode,
   imageCDNState,
+  generateTypeName,
 } from "./normalize"
 
-const {
+import {
   handleReferences,
   handleWebhookUpdate,
   createNodeIfItDoesNotExist,
   handleDeletedNode,
   drupalCreateNodeManifest,
   getExtendedFileNodeData,
-} = require(`./utils`)
-
+} from "./utils"
 const imageCdnDocs = `https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal#readme`
 
 const agent = {
@@ -243,6 +243,7 @@ ${JSON.stringify(webhookBody, null, 4)}`
             createNodeId,
             createContentDigest,
             entityReferenceRevisions,
+            pluginOptions,
           })
           reporter.log(`Deleted node: ${deletedNode.id}`)
         }
@@ -264,6 +265,7 @@ ${JSON.stringify(webhookBody, null, 4)}`
           createContentDigest,
           getNode,
           reporter,
+          pluginOptions,
         })
       }
 
@@ -278,8 +280,6 @@ ${JSON.stringify(webhookBody, null, 4)}`
             getCache,
             getNode,
             reporter,
-            store,
-            languageConfig,
           },
           pluginOptions
         )
@@ -402,6 +402,7 @@ ${JSON.stringify(webhookBody, null, 4)}`
                 createContentDigest,
                 getNode,
                 reporter,
+                pluginOptions,
               })
             }
           }
@@ -416,6 +417,7 @@ ${JSON.stringify(webhookBody, null, 4)}`
                 createNodeId,
                 createContentDigest,
                 entityReferenceRevisions,
+                pluginOptions,
               })
             } else {
               // The data could be a single Drupal entity or an array of Drupal
@@ -436,8 +438,6 @@ ${JSON.stringify(webhookBody, null, 4)}`
                     getCache,
                     getNode,
                     reporter,
-                    store,
-                    languageConfig,
                   },
                   pluginOptions
                 )
@@ -773,6 +773,7 @@ ${JSON.stringify(webhookBody, null, 4)}`
       createNodeId,
       cache,
       entityReferenceRevisions,
+      pluginOptions,
     })
   }
 
@@ -782,7 +783,9 @@ ${JSON.stringify(webhookBody, null, 4)}`
     reporter.info(`Downloading remote files from Drupal`)
 
     // Download all files (await for each pool to complete to fix concurrency issues)
-    const fileNodes = [...nodes.values()].filter(isFileNode)
+    const fileNodes = [...nodes.values()].filter(node =>
+      isFileNode(node, pluginOptions.typePrefix)
+    )
 
     if (fileNodes.length) {
       const downloadingFilesActivity = reporter.activityTimer(
@@ -795,12 +798,10 @@ ${JSON.stringify(webhookBody, null, 4)}`
           await downloadFile(
             {
               node,
-              store,
               cache,
               createNode,
               createNodeId,
               getCache,
-              reporter,
             },
             pluginOptions
           )
@@ -833,7 +834,6 @@ exports.onCreateDevServer = (
     createNodeId,
     getNode,
     actions,
-    store,
     cache,
     createContentDigest,
     getCache,
@@ -873,7 +873,6 @@ exports.onCreateDevServer = (
             getCache,
             getNode,
             reporter,
-            store,
           },
           pluginOptions
         )
@@ -913,6 +912,9 @@ exports.pluginOptionsSchema = ({ Joi }) =>
     disallowedLinkTypes: Joi.array().items(Joi.string()),
     skipFileDownloads: Joi.boolean(),
     imageCDN: Joi.boolean().default(true),
+    typePrefix: Joi.string()
+      .description(`Prefix for Drupal node types`)
+      .default(``),
     fastBuilds: Joi.boolean(),
     entityReferenceRevisions: Joi.array().items(Joi.string()),
     secret: Joi.string().description(
@@ -950,10 +952,22 @@ exports.createSchemaCustomization = (
   if (pluginOptions.skipFileDownloads && pluginOptions.imageCDN) {
     actions.createTypes([
       // polyfill so image CDN works on older versions of Gatsby
+      // this type is merged in with the inferred file__file and files types, adding Image CDN support via the gatsbyImage GraphQL field. The `RemoteFile` interface as well as the polyfill above are what add the gatsbyImage field.
       addRemoteFilePolyfillInterface(
-        // this type is merged in with the inferred file__file type, adding Image CDN support via the gatsbyImage GraphQL field. The `RemoteFile` interface as well as the polyfill above are what add the gatsbyImage field.
         schema.buildObjectType({
-          name: `file__file`,
+          name: generateTypeName(`file--file`, pluginOptions.typePrefix),
+          fields: {},
+          interfaces: [`Node`, `RemoteFile`],
+        }),
+        {
+          schema,
+          actions,
+          store,
+        }
+      ),
+      addRemoteFilePolyfillInterface(
+        schema.buildObjectType({
+          name: generateTypeName(`files`, pluginOptions.typePrefix),
           fields: {},
           interfaces: [`Node`, `RemoteFile`],
         }),