+++ packages/gatsby-source-drupal/src/normalize.ts
@@ -1,6 +1,7 @@
 const { URL } = require(`url`)
 const { createRemoteFileNode } = require(`gatsby-source-filesystem`)
 const path = require(`path`)
+const { capitalize } = require(`lodash`)
 const probeImageSize = require(`probe-image-size`)
 
 import { getOptions } from "./plugin-options"
@@ -33,17 +34,20 @@ const getGatsbyImageCdnFields = async ({
     return {}
   }
 
-  const isFile = isFileNode({
-    internal: {
-      type,
+  const isFile = isFileNode(
+    {
+      internal: {
+        type,
+      },
     },
-  })
+    pluginOptions.typePrefix
+  )
 
   if (!isFile) {
     return {}
   }
 
-  const mimeType = node.attributes.filemime
+  const mimeType = node.attributes.filemime || node.attributes.mimetype
   const { filename } = node.attributes
 
   if (!mimeType || !filename) {
@@ -139,6 +143,14 @@ const getGatsbyImageCdnFields = async ({
   return {}
 }
 
+export const generateTypeName = (type: string, typePrefix = ``) => {
+  const prefixed = typePrefix
+    ? `${capitalize(typePrefix)}${capitalize(type)}`
+    : type
+
+  return prefixed.replace(/-|__|:|\.|\s/g, `_`)
+}
+
 export const nodeFromData = async (
   datum,
   createNodeId,
@@ -154,7 +166,9 @@ export const nodeFromData = async (
     typeof attributeId !== `undefined` ? { _attributes_id: attributeId } : {}
 
   const langcode = attributes.langcode || `und`
-  const type = datum.type.replace(/-|__|:|\.|\s/g, `_`)
+  const { typePrefix = `` } = pluginOptions
+
+  const type = generateTypeName(datum.type, typePrefix)
 
   const gatsbyImageCdnFields = await getGatsbyImageCdnFields({
     node: datum,
@@ -164,13 +178,14 @@ export const nodeFromData = async (
     reporter,
   })
 
-  const versionedId = createNodeIdWithVersion(
-    datum.id,
-    datum.type,
+  const versionedId = createNodeIdWithVersion({
+    id: datum.id,
+    type: datum.type,
     langcode,
-    attributes.drupal_internal__revision_id,
-    entityReferenceRevisions
-  )
+    revisionId: attributes.drupal_internal__revision_id,
+    entityReferenceRevisions,
+    typePrefix,
+  })
 
   const gatsbyId = createNodeId(versionedId)
 
@@ -191,18 +206,29 @@ export const nodeFromData = async (
   }
 }
 
-const isEntityReferenceRevision = (type, entityReferenceRevisions = []) =>
+const isEntityReferenceRevision = (
+  type,
+  entityReferenceRevisions: Array<string> = []
+) =>
   entityReferenceRevisions.findIndex(
     revisionType => type.indexOf(revisionType) === 0
   ) !== -1
 
-export const createNodeIdWithVersion = (
-  id: string,
-  type: string,
-  langcode: string,
-  revisionId: string,
-  entityReferenceRevisions = []
-) => {
+export const createNodeIdWithVersion = ({
+  id,
+  type,
+  langcode,
+  revisionId,
+  entityReferenceRevisions = [],
+  typePrefix,
+}: {
+  id: string
+  type: string
+  langcode: string
+  revisionId: string
+  entityReferenceRevisions: Array<string>
+  typePrefix: string
+}) => {
   const options = getOptions()
 
   // Fallback to default language for entities that don't translate.
@@ -245,12 +271,23 @@ export const createNodeIdWithVersion = (
     ? `${langcodeNormalized}.${id}.${revisionId || 0}`
     : `${langcodeNormalized}.${id}`
 
-  return idVersion
+  return typePrefix ? `${typePrefix}.${idVersion}` : idVersion
 }
 
-export const isFileNode = node => {
-  const type = node?.internal?.type
-  return type === `files` || type === `file__file`
+const fileNodeTypes = new Map<string, Set<string>>()
+
+export const isFileNode = (node, typePrefix = ``) => {
+  if (!fileNodeTypes.has(typePrefix)) {
+    fileNodeTypes.set(
+      typePrefix,
+      new Set<string>([
+        generateTypeName(`files`, typePrefix),
+        generateTypeName(`file--file`, typePrefix),
+      ])
+    )
+  }
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  return fileNodeTypes.get(typePrefix)!.has(node?.internal?.type)
 }
 
 const getFileUrl = (node, baseUrl) => {
@@ -269,41 +306,42 @@ const getFileUrl = (node, baseUrl) => {
 
 export const downloadFile = async (
   { node, cache, createNode, createNodeId, getCache },
-  { basicAuth, baseUrl }
+  { basicAuth, baseUrl, typePrefix }: Record<string, any>
 ) => {
   // handle file downloads
-  if (isFileNode(node)) {
-    let fileType
+  if (!isFileNode(node, typePrefix)) {
+    return
+  }
+  let fileType
 
-    if (typeof node.uri === `object`) {
-      // get file type from uri prefix ("S3:", "public:", etc.)
-      const uriPrefix = node.uri.value.match(/^\w*:/)
-      fileType = uriPrefix ? uriPrefix[0] : null
-    }
+  if (typeof node.uri === `object`) {
+    // get file type from uri prefix ("S3:", "public:", etc.)
+    const uriPrefix = node.uri.value.match(/^\w*:/)
+    fileType = uriPrefix ? uriPrefix[0] : null
+  }
 
-    const url = getFileUrl(node, baseUrl)
+  const url = getFileUrl(node, baseUrl)
 
-    // If we have basicAuth credentials, add them to the request.
-    const basicAuthFileSystems = [`public:`, `private:`, `temporary:`]
-    const auth =
-      typeof basicAuth === `object` && basicAuthFileSystems.includes(fileType)
-        ? {
-            htaccess_user: basicAuth.username,
-            htaccess_pass: basicAuth.password,
-          }
-        : {}
-    const fileNode = await createRemoteFileNode({
-      url: url.href,
-      name: path.parse(decodeURIComponent(url.pathname)).name,
-      cache,
-      createNode,
-      createNodeId,
-      getCache,
-      parentNodeId: node.id,
-      auth,
-    })
-    if (fileNode) {
-      node.localFile___NODE = fileNode.id
-    }
+  // If we have basicAuth credentials, add them to the request.
+  const basicAuthFileSystems = [`public:`, `private:`, `temporary:`]
+  const auth =
+    typeof basicAuth === `object` && basicAuthFileSystems.includes(fileType)
+      ? {
+          htaccess_user: basicAuth.username,
+          htaccess_pass: basicAuth.password,
+        }
+      : {}
+  const fileNode = await createRemoteFileNode({
+    url: url.href,
+    name: path.parse(decodeURIComponent(url.pathname)).name,
+    cache,
+    createNode,
+    createNodeId,
+    getCache,
+    parentNodeId: node.id,
+    auth,
+  })
+  if (fileNode) {
+    node.localFile___NODE = fileNode.id
   }
 }