+++ packages/gatsby-plugin-feed/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-feed@5.9.0/packages/gatsby-plugin-feed) (2023-04-18)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)
+
+#### Bug Fixes
+
+- update dependency fs-extra to ^11.1.1 [#37827](https://github.com/gatsbyjs/gatsby/issues/37827) ([3e9a590](https://github.com/gatsbyjs/gatsby/commit/3e9a590fdc0b4cdbc763b6a6e3ca3e983ac8231a))
+
+#### Chores
+
+- Update README with clearer instructions [#37930](https://github.com/gatsbyjs/gatsby/issues/37930) ([4a1b2ce](https://github.com/gatsbyjs/gatsby/commit/4a1b2ce278d08bd6b473e891dab588efc51c9f63))
+
 ## [5.8.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-feed@5.8.0/packages/gatsby-plugin-feed) (2023-03-21)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.8)