+++ packages/gatsby-remark-images-contentful/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-remark-images-contentful@6.9.0/packages/gatsby-remark-images-contentful) (2023-04-18)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)
+
+**Note:** Version bump only for package gatsby-remark-images-contentful
+
 ## [6.8.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-remark-images-contentful@6.8.0/packages/gatsby-remark-images-contentful) (2023-03-21)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.8)